requirements.txt
asgiref==3.8.1
Django==4.2.11
django-bootstrap5==22.2
django-debug-toolbar==4.3.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
